// src/models/HistoricoManutencao.ts
type HistoricoManutencao = {
    CD_TABELA?: number;
    DT_FIM_MANUTENCAO?: string;
    DT_INICIO_MANUTENCAO?: string;
    NR_OS?: number;
    NR_SEQUENCIA?: number;
    NR_SEQ_LOGICA?: number;
    SG_SETOR?: string;
    SG_UNIDADE?: string;
    TP_REFERENCIA?: string;
}

export default HistoricoManutencao;